folders and CSS modules
moved components into folders and added CSS modules for each component
