--- my-todo-frontend/src/Content.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import InputTextBox from './InputTextBox';
-
-function Content() {
-    return (
-        <div id="Content">
-            <InputTextBox properties={{
-                width: 550,
-                placeholder: 'Enter todo name...'
-            }} />
-        </div>
-    );
-}
-
-export default Content;
