+++ my-todo-frontend/src/components/Header.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
+import styles from '../styles/Header.module.css';
 
 function Header() {
     return (
-        <header id="Header">
+        <header className={styles.Header}>
             <h4>my-todo-frontend</h4>
         </header>
     );
