+++ my-todo-frontend/src/index.css
@@ -15,3 +15,7 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
+
+a {
+    color: #61dafb;
+}
