+++ my-todo-frontend/src/styles/Content.module.css
@@ -1,0 +1,4 @@
+.Content {
+    display: flex;
+    justify-content: center;
+}
