+++ my-todo-frontend/src/styles/Header.module.css
@@ -1,0 +1,4 @@
+.Header {
+    display: flex;
+    justify-content: center;
+}
