+++ my-todo-frontend/src/styles/InputTextBox.module.css
@@ -1,26 +1,3 @@
-#App {
-    display: flex;
-    flex-flow: column wrap;
-    height: 100vh;
-    background-color: #282c34;
-    color: white;
-    font-size: calc(10px + 2vmin);
-}
-
-#Header {
-    display: flex;
-    justify-content: center;
-}
-
-#Content {
-    display: flex;
-    justify-content: center;
-}
-
-a {
-    color: #61dafb;
-}
-
 .InputTextBox {
     height: 56px;
     border-radius: 4px;
@@ -28,7 +5,7 @@ a {
     transition: 0.3s background-color ease-in-out, 0.3s box-shadow ease-in-out;
 }
 
-.InputTextBox input {  
+.InputTextBox input {
     height: 56px;
     padding: 0px 16px;
     border: none;
